+++ src/pathsAndNames.ts
@@ -5,11 +5,10 @@ import { timestampOffsetMinutes } from './consts';
   media v1
    |- $username1
    |-|- $YYYY-MM
-   |-|-|- $DD $author $mediaCode .ext // .jpg | .mp4
-   |-|-|- $DD $author $mediaCode // carousel
-   |-|-|-|- $carouselCounter .ext // starts from 1. .jpg | .mp4
-    - $username2
-      - ...
+   |-|-|- [$author] $DD $mediaCode .ext // .jpg | .mp4
+   |-|-|- [$author] $DD $mediaCode $carouselCounter .ext // starts from 1. .jpg | .mp4
+   |- $username2
+   |-|- ...
 */
 
 
@@ -19,14 +18,14 @@ import { timestampOffsetMinutes } from './consts';
 
 /** Naming it media vX so we can change the version for breaking changes without
  * messing existing local data. */
-export const mediaDirName = 'media v1';
+export const mediaDirName = 'media v2';
 
 export function getMediaUserDirRelPath({ username, forcePosix = false }: {
   username: string;
   /** Set forcePosix: true for globby. Else leave false for automatic path joining */
   forcePosix?: boolean;
 }): string {
-  return forcePosix ? Path.posix.join('media v1', username) : Path.join('media v1', username);
+  return forcePosix ? Path.posix.join(mediaDirName, username) : Path.join(mediaDirName, username);
 }
 
 export function getMediaUserDateDirRelPath({ username, date, forcePosix = false }: {
@@ -49,27 +48,28 @@ export function getDirDateName({ date }: {date: Date}): string {
 }
 
 /** Without extension. */
-export function getFilename({ date, code, author }: {
-  date: Date, code: string, author: string
+export function getFilename({ date, code, author, carouselCounter }: {
+  date: Date; code: string; author: string; carouselCounter?: number
 }): string {
   // https://stackoverflow.com/a/29774197/10247962
   const offsettedDate = new Date(date.getTime() - (timestampOffsetMinutes * 60 * 1000));
   // const YYYYmmDD = offsettedDate.toISOString().split('T')[0]!;
   // const HHmmSS = offsettedDate.toISOString().split('T')[1]!.split('.')[0]!;
   const dd = String(offsettedDate.getDate()).padStart(2, '0');
-  const filename = `${dd} [${author}] ${code}`;
+  const filename = `[${author}] ${dd} ${code}` + (carouselCounter === undefined ? '' : ` ${carouselCounter}`);
 
   return filename;
 }
 
-/** Will call getFilename(). */
-export function getCarouselDirName({ date, code, author }: {
-  date: Date, code: string; author: string;
-}): string {
-  return getFilename({ date, code: code, author });
-}
+// The client doesn't want to have a directory no more.
+// /** Will call getFilename(). */
+// export function getCarouselDirName({ date, code, author }: {
+//   date: Date, code: string; author: string;
+// }): string {
+//   return getFilename({ date, code: code, author });
+// }
 
-/** Will just return the carouselCounter as string. Won't include the file extension. */
-export function getCarouselFilename({ carouselCounter }: {carouselCounter: number}): string {
-  return String(carouselCounter);
-}
+// /** Will just return the carouselCounter as string. Won't include the file extension. */
+// export function getCarouselFilename({ carouselCounter }: {carouselCounter: number}): string {
+//   return String(carouselCounter);
+// }