+++ CHANGELOG.md
@@ -11,6 +11,14 @@
 *
 -->
 
+
+# 1.2.0 (2021-06-01) - Client requirements changed
+### Changed
+* Filename changed from `$day [$author] $code` to `[$author] $day $code`
+* Carousels are no longer directories
+* `media v1` to `media v2`
+
+
 # 1.1.0 (2021-05-15)
 ### Added
 * Press any key to exit program (good for windows)